2D game base setup
I've added a quick 2d game set up, but it needs a more fleshed out world and a way to get the last code. You can interact with the arcade and the game should play. I still need to add ui things for the 2d games, but progress is going well.